fix: moved tests with mock Persona to their own test file, so they don't break tests relying on real Persona component being used
there is probably some magic Jest command that would make this unnecessary but I gave up looking for it
