+++ packages/mobile/src/account/LinkBankAccountScreen.integration.test.tsx
@@ -1,7 +1,7 @@
 import { fireEvent, render, waitFor } from '@testing-library/react-native'
 import * as React from 'react'
 import 'react-native'
-import LinkBankAccountScreen, { StepOne } from 'src/account/LinkBankAccountScreen'
+import LinkBankAccountScreen from 'src/account/LinkBankAccountScreen'
 import { KycStatus } from 'src/account/reducer'
 import { navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
@@ -10,21 +10,8 @@ import { createMockStore } from 'test/utils'
 import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
 import { CICOEvents } from 'src/analytics/Events'
 import { mockAccount } from 'test/values'
-import PersonaButton from './Persona'
-import Button from '@celo/react-components/components/Button'
 
 jest.mock('src/analytics/ValoraAnalytics')
-
-let personaButtonSuccessCallback: (() => any) | undefined // using this to simulate Persona success at any arbitrary time
-const MockPersona = ({ onSuccess, onPress }: { onSuccess: () => any; onPress: () => any }) => {
-  personaButtonSuccessCallback = onSuccess
-  return <Button onPress={onPress} text="test persona button" testID="PersonaButton" />
-}
-jest.mock('./Persona', () => ({
-  __esModule: true,
-  namedExport: jest.fn(),
-  default: jest.fn(),
-}))
 
 const FAKE_TEMPLATE_ID = 'fake template id'
 jest.mock('react-native-persona')
@@ -37,7 +24,7 @@ jest.mock('src/in-house-liquidity', () => ({
   createPersonaAccount: jest.fn(() => mockResponse),
 }))
 
-describe('LinkBankAccountScreen', () => {
+describe('LinkBankAccountScreen: integration tests (using real Persona component, for instance)', () => {
   beforeEach(() => {
     jest.useRealTimers()
     jest.clearAllMocks()
@@ -96,31 +83,5 @@ describe('LinkBankAccountScreen', () => {
     fireEvent.press(getByTestId('PersonaButton'))
     await waitFor(() => getByText('linkBankAccountScreen.verifying.title'))
     expect(ValoraAnalytics.track).toHaveBeenCalledWith(CICOEvents.persona_kyc_start)
-  })
-
-  describe('Tests with mock Persona button', () => {
-    beforeAll(() => {
-      //@ts-ignore . my IDE complains about this, though jest allows it
-      PersonaButton.mockImplementation(MockPersona)
-    })
-    afterEach(() => {
-      personaButtonSuccessCallback = undefined
-    })
-    afterAll(() => {
-      jest.clearAllMocks()
-    })
-    it('switches to completed state when persona onSuccess callback is called and kyc exists', async () => {
-      const { getByText, getByTestId, rerender } = render(<StepOne kycStatus={undefined} />)
-      await waitFor(() => expect(getByTestId('PersonaButton')).not.toBeDisabled())
-
-      await fireEvent.press(getByTestId('PersonaButton'))
-      await waitFor(() => getByText('linkBankAccountScreen.verifying.title'))
-
-      rerender(<StepOne kycStatus={KycStatus.Approved} />) // simulating the redux store being updated to reflect new KYC status. in a production flow Persona calls an IHL webhook, which updates firebase, which triggers a redux store update
-
-      expect(personaButtonSuccessCallback).toBeTruthy()
-      personaButtonSuccessCallback?.()
-      await waitFor(() => getByText('linkBankAccountScreen.completed.title')) // spinny wheel gone
-    })
   })
 })
